requirements.txt
torch
torchvision
torchaudio
transformers
pillow
easyocr
langchain
streamlit
sentencepiece
accelerate